+++ docs/CHANGELOG.md
@@ -1,6 +1,13 @@
 CHANGELOG
 =========
 
+Version 1.0.3-dev
+-----------------
+
+* remove variable echo, dir_template and dir_cache
+* add config array
+* add setConfig() getConfig()
+
 Version 1.0.2
 -------------